+++ previous-versions/commerce/src/codatcommerce/models/operations/get_company_info.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetCompanyInfoRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     
 
 
@@ -21,11 +23,14 @@ class GetCompanyInfoRequest:
 @dataclasses.dataclass
 class GetCompanyInfoResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     company_info: Optional[shared_companyinfo.CompanyInfo] = dataclasses.field(default=None)
     r"""OK"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""