+++ previous-versions/commerce/src/codatcommerce/models/shared/taxcomponentallocation.py
@@ -5,6 +5,7 @@ import dataclasses
 from ..shared import taxcomponentref as shared_taxcomponentref
 from codatcommerce import utils
 from dataclasses_json import Undefined, dataclass_json
+from decimal import Decimal
 from typing import Optional
 
 
@@ -12,7 +13,7 @@ from typing import Optional
 
 @dataclasses.dataclass
 class TaxComponentAllocation:
-    rate: Optional[float] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('rate'), 'exclude': lambda f: f is None }})
+    rate: Optional[Decimal] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('rate'), 'encoder': utils.decimalencoder(True, False), 'decoder': utils.decimaldecoder }})
     r"""Tax amount on order line sale as available from source commerce platform."""
     tax_component_ref: Optional[shared_taxcomponentref.TaxComponentRef] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('taxComponentRef'), 'exclude': lambda f: f is None }})
     r"""Taxes rates reference object depending on the rates being available on source commerce package."""